Add Magnifier icon
The Magnifier and MagnifierSmall icons only render an <svg> element.  It
is adviced to only ever use these icons as they come but it's possible
to pass props to override the default attributes on the <svg> element.

There is also a convenience export that allowes the following:

```TypeScript
import * as Icon from '../path/to/icons';
```

…which  allows all available icons to be used like this:

```TSX
// ...
<Icon.Magnifier />
// ...
```